restructured example, extended ping registration script
